Remove redundant volume loop for media players
Volume for all media players depends on the mediaplayer_volume convar. For players with 3d audio enabled, this loop causes a "blip" in volume even when the player is out of proximity. Additionally this causes a temporary dip in volume levels > 50% when opening the sidebar because MP.VolumeSlider will default to 0.5 on initialization.